+++ utils/AnalyzerUtils/CrossTest.py
@@ -12,6 +12,28 @@ class CrossTest:
     def __init__(self, tfa):
         self.tfa = tfa
 
+    def _get_df_primary_secondary(self, df_type: str):
+        if df_type not in ['taxa-func', 'func-taxa', 'taxa', 'func', 'peptide', 'protein']:
+            raise ValueError("df_type must be in ['taxa-func', 'func-taxa', 'taxa', 'func', 'peptide']")
+        
+        df, primary, secondary = None, None, None
+        
+        if df_type == 'taxa-func':
+            df, primary, secondary = self.tfa.taxa_func_df, 'Taxon', self.tfa.func_name
+        elif df_type == 'func-taxa':
+            df, primary, secondary = self.tfa.func_taxa_df, self.tfa.func_name, 'Taxon'
+        elif df_type == 'taxa':
+            df, primary = self.tfa.taxa_df, 'Taxon'
+        elif df_type == 'func':
+            df, primary = self.tfa.func_df, self.tfa.func_name
+        elif df_type == 'peptide':
+            df, primary = self.tfa.peptide_df, self.tfa.peptide_col_name
+        elif df_type == 'protein':
+            df, primary = self.tfa.protein_df, self.tfa.protein_col_name
+            
+        return df, primary, secondary
+            
+
 
     def get_stats_anova(self, group_list: list = None, df_type:str = 'taxa-func') -> pd.DataFrame:
         group_list_all = sorted(set(self.tfa.get_meta_list(self.tfa.meta_name)))
@@ -26,53 +48,40 @@ class CrossTest:
 
         all_sample_list = [sample for group in group_list for sample in self.tfa.get_sample_list_in_a_group(group)]
 
-        secondary = None # give a default value to prevent error when print
-        if df_type in ['taxa-func', 'func-taxa', 'taxa', 'func', 'peptide']:
-            if df_type == 'taxa-func':
-                df, primary, secondary = self.tfa.taxa_func_df, 'Taxon', self.tfa.func_name
-            elif df_type == 'func-taxa':
-                df, primary, secondary = self.tfa.func_taxa_df, self.tfa.func_name, 'Taxon'
-            elif df_type == 'taxa':
-                df, primary = self.tfa.taxa_df, 'Taxon'
-            elif df_type == 'func':
-                df, primary = self.tfa.func_df, self.tfa.func_name
-            elif df_type == 'peptide':
-                df, primary = self.tfa.peptide_df, self.tfa.peptide_col_name
-            else:
-                raise ValueError("df_type must be in ['taxa-func', 'func-taxa', 'taxa', 'func', 'peptide']")
-            
-            if secondary is not None:
-                print(f"ANOVA test for {primary}-{secondary} in {group_list}")
-            else:
-                print(f"ANOVA test for {primary} in {group_list}")
+        df, primary, secondary = self._get_df_primary_secondary(df_type)
+        
+        if secondary is not None:
+            print(f"ANOVA test for {primary}-{secondary} in {group_list}")
+        else:
+            print(f"ANOVA test for {primary} in {group_list}")
 
-            res = {primary: [], "P-value": [], "f-statistic": []}
+        res = {primary: [], "P-value": [], "f-statistic": []}
+        if df_type in ['taxa-func', 'func-taxa']:
+            res[secondary] = []
+
+        for row in tqdm(df.iterrows(), total=len(df)):
+            primary_value = row[0]
             if df_type in ['taxa-func', 'func-taxa']:
-                res[secondary] = []
+                primary_value = row[0][0]
+                secondary_value = row[0][1]
+                res[secondary].append(secondary_value)
 
-            for row in tqdm(df.iterrows(), total=len(df)):
-                primary_value = row[0]
-                if df_type in ['taxa-func', 'func-taxa']:
-                    primary_value = row[0][0]
-                    secondary_value = row[0][1]
-                    res[secondary].append(secondary_value)
+            res[primary].append(primary_value)
 
-                res[primary].append(primary_value)
+            list_for_anova = [row[1][self.tfa.get_sample_list_in_a_group(group)].to_list() for group in group_list]
 
-                list_for_anova = [row[1][self.tfa.get_sample_list_in_a_group(group)].to_list() for group in group_list]
+            f, p = f_oneway(*list_for_anova)
+            res["P-value"].append(p)
+            res["f-statistic"].append(f)
 
-                f, p = f_oneway(*list_for_anova)
-                res["P-value"].append(p)
-                res["f-statistic"].append(f)
-
-            res = pd.DataFrame(res)
-            on_values = [primary]
-            if df_type in ['taxa-func', 'func-taxa']:
-                on_values.append(secondary)
-            res_all = pd.merge(df, res, on=on_values)
-            res_all.index = df.index
-            res_all = res_all[['P-value', 'f-statistic'] + all_sample_list]
-            return res_all
+        res = pd.DataFrame(res)
+        on_values = [primary]
+        if df_type in ['taxa-func', 'func-taxa']:
+            on_values.append(secondary)
+        res_all = pd.merge(df, res, on=on_values)
+        res_all.index = df.index
+        res_all = res_all[['P-value', 'f-statistic'] + all_sample_list]
+        return res_all
         
     def get_stats_ttest(self, group_list: list = None, df_type: str = 'taxa-func') -> pd.DataFrame:
 
@@ -87,19 +96,7 @@ class CrossTest:
 
         all_sample_list = [sample for group in group_list for sample in self.tfa.get_sample_list_in_a_group(group)]
 
-        secondary = None # give a default value to prevent error when print
-        if df_type == 'taxa-func':
-            df, primary, secondary = self.tfa.taxa_func_df, 'Taxon', self.tfa.func_name
-        elif df_type == 'func-taxa':
-            df, primary, secondary = self.tfa.func_taxa_df, self.tfa.func_name, 'Taxon'
-        elif df_type == 'taxa':
-            df, primary = self.tfa.taxa_df, 'Taxon'
-        elif df_type == 'func':
-            df, primary = self.tfa.func_df, self.tfa.func_name
-        elif df_type == 'peptide':
-            df, primary = self.tfa.peptide_df, self.tfa.peptide_col_name
-        else:
-            raise ValueError("df_type must be in ['taxa-func', 'func-taxa', 'taxa', 'func', 'peptide']")
+        df, primary, secondary = self._get_df_primary_secondary(df_type)
 
         res = {primary: [], "P-value": [], "t-statistic": []}
         
@@ -158,19 +155,7 @@ class CrossTest:
             group_list = sorted(set(group_list))
             
    
-        secondary_index = None # give a default value to prevent error when print
-        if df_type == 'taxa-func':
-            df, primary_index, secondary_index = self.tfa.taxa_func_df, 'Taxon', self.tfa.func_name
-        elif df_type == 'func-taxa':
-            df, primary_index, secondary_index = self.tfa.func_taxa_df, self.tfa.func_name, 'Taxon'
-        elif df_type == 'taxa':
-            df, primary_index = self.tfa.taxa_df, 'Taxon'
-        elif df_type == 'func':
-            df, primary_index = self.tfa.func_df, self.tfa.func_name
-        elif df_type == 'peptide':
-            df, primary_index = self.tfa.peptide_df, self.tfa.peptide_col_name
-        else:
-            raise ValueError("df_type must be in ['taxa-func', 'func-taxa', 'taxa', 'func', 'peptide']")
+        df, primary_index, secondary_index = self._get_df_primary_secondary(df_type)
 
         res_dict = {primary_index: [], "p_value": [], "t_statistic": []}